bas-io: IStreamResource.getLength returns long instead of Long.

+++ bas/layer-2/bas-io/src/main/java/net/bodz/bas/io/res/builtin/ByteArrayResource.java
@@ -58,8 +58,8 @@ public class ByteArrayResource
     }
 
     @Override
-    public Long getLength() {
-        return (long) (end - start);
+    public long getLength() {
+        return end - start;
     }
 
     @Override
