+++ bas/layer-2/bas-io/src/main/java/net/bodz/bas/io/res/builtin/CharArrayResource.java
@@ -51,8 +51,8 @@ public class CharArrayResource
     }
 
     @Override
-    public Long getLength() {
-        return (long) (end - start);
+    public long getLength() {
+        return end - start;
     }
 
     @Override
